A few more edits... now I'm finished
I have left a few comments in, they appear in italics in the pdf and
inside _(LM … )_ in the .Rmd so are pink. I have commented out Len’s
comments as I have tried to address them but they are still in the .Rmd
file in <!— —> and appear in blue. Actually colours may vary depending
on your personal setup.